refactor(collections): simplify dynamic query assembly
