+++ src/main/java/com/researchspace/dao/query/CollectionQueryExecutor.java
@@ -6,12 +6,12 @@ import com.blazebit.persistence.MultipleSubqueryInitiator;
 import com.blazebit.persistence.PagedList;
 import com.researchspace.dao.query.RsqlCollectionQuery.Predicate;
 import com.researchspace.model.collection.CollectionDescription;
-import com.researchspace.model.collection.CollectionDescription.Sort;
 import com.researchspace.model.collection.FilterExpression;
 import com.researchspace.model.collection.QueryConstraint;
 import com.researchspace.model.collection.RelationshipReadAccess;
 import com.researchspace.model.collection.ResourcePage;
 import com.researchspace.model.collection.ResourceRequest;
+import com.researchspace.model.collection.Sort;
 import jakarta.persistence.LockModeType;
 import java.util.List;
 import java.util.Objects;
