+++ src/main/java/com/researchspace/maintenance/model/ApiV2MaintenanceResource.java
@@ -8,10 +8,10 @@ import com.researchspace.model.collection.AccessResult;
 import com.researchspace.model.collection.ApiV2ResourceDefinition;
 import com.researchspace.model.collection.ApiV2ResourceField;
 import com.researchspace.model.collection.CollectionDescription;
-import com.researchspace.model.collection.CollectionDescription.Operator;
-import com.researchspace.model.collection.CollectionDescription.Sort;
 import com.researchspace.model.collection.CollectionMutationLimits;
 import com.researchspace.model.collection.FilterExpression;
+import com.researchspace.model.collection.Operator;
+import com.researchspace.model.collection.Sort;
 import java.util.Date;
 import java.util.List;
 import java.util.Set;
