+++ src/main/java/com/researchspace/model/collection/AccessPolicySchema.java
@@ -1,0 +1,8 @@
+package com.researchspace.model.collection;
+
+public record AccessPolicySchema(
+    AccessDocumentation readAccess,
+    AccessDocumentation createAccess,
+    AccessDocumentation updateAccess,
+    AccessDocumentation deleteAccess,
+    AccessDocumentation softDeleteAccess) {}
