+++ src/main/java/com/researchspace/model/collection/ApiV2UserResource.java
@@ -3,7 +3,6 @@ package com.researchspace.model.collection;
 import static com.researchspace.model.collection.ApiV2ResourceField.AccessPreset.NEVER;
 
 import com.researchspace.model.User;
-import com.researchspace.model.collection.CollectionDescription.Sort;
 import java.util.List;
 
 /**
