+++ src/main/java/com/researchspace/model/collection/FilterExpression.java
@@ -1,6 +1,5 @@
 package com.researchspace.model.collection;
 
-import com.researchspace.model.collection.CollectionDescription.Operator;
 import java.util.List;
 import java.util.Objects;
