+++ src/main/java/com/researchspace/model/collection/FilterSchema.java
@@ -1,0 +1,5 @@
+package com.researchspace.model.collection;
+
+import java.util.Set;
+
+public record FilterSchema(String selector, Set<Operator> operators, boolean supportsWildcards) {}
