+++ src/main/java/com/researchspace/model/collection/Operator.java
@@ -1,0 +1,15 @@
+package com.researchspace.model.collection;
+
+public enum Operator {
+  EQUAL,
+  NOT_EQUAL,
+  GREATER_THAN,
+  GREATER_THAN_OR_EQUAL,
+  LESS_THAN,
+  LESS_THAN_OR_EQUAL,
+  IN,
+  NOT_IN,
+  CONTAINS,
+  LIKE,
+  EXISTS
+}
