+++ src/main/java/com/researchspace/model/collection/RuntimeFieldDefinition.java
@@ -1,6 +1,5 @@
 package com.researchspace.model.collection;
 
-import com.researchspace.model.collection.CollectionDescription.Operator;
 import java.util.List;
 import java.util.Objects;
 import java.util.Set;
