+++ src/main/java/com/researchspace/model/collection/Sort.java
@@ -1,0 +1,3 @@
+package com.researchspace.model.collection;
+
+public record Sort(String field, boolean ascending) {}
