+++ src/main/java/com/researchspace/model/collection/WriteOperation.java
@@ -1,0 +1,6 @@
+package com.researchspace.model.collection;
+
+public enum WriteOperation {
+  CREATE,
+  UPDATE
+}
