+++ src/main/java/com/researchspace/service/inventory/InstrumentReadAccess.java
@@ -11,8 +11,8 @@ import com.researchspace.model.collection.AccessDocumentation;
 import com.researchspace.model.collection.AccessFunction;
 import com.researchspace.model.collection.AccessPolicy;
 import com.researchspace.model.collection.AccessResult;
-import com.researchspace.model.collection.CollectionDescription.Operator;
 import com.researchspace.model.collection.FilterExpression;
+import com.researchspace.model.collection.Operator;
 import com.researchspace.model.inventory.InventoryRecord.InventorySharingMode;
 import java.util.ArrayList;
 import java.util.List;
