+++ src/test/java/com/researchspace/dao/InstrumentDaoCollectionTest.java
@@ -7,12 +7,12 @@ import com.researchspace.model.User;
 import com.researchspace.model.collection.AccessContext;
 import com.researchspace.model.collection.AccessContext.Operation;
 import com.researchspace.model.collection.AccessResult;
-import com.researchspace.model.collection.CollectionDescription.Operator;
-import com.researchspace.model.collection.CollectionDescription.Sort;
 import com.researchspace.model.collection.FieldSelection;
 import com.researchspace.model.collection.FilterExpression;
 import com.researchspace.model.collection.IncludeTree;
+import com.researchspace.model.collection.Operator;
 import com.researchspace.model.collection.ResourceRequest;
+import com.researchspace.model.collection.Sort;
 import com.researchspace.model.inventory.Instrument;
 import com.researchspace.service.inventory.InstrumentReadAccess;
 import com.researchspace.testutils.SpringTransactionalTest;
