+++ src/test/java/com/researchspace/dao/hibernate/InventoryReadRestrictionTest.java
@@ -12,8 +12,8 @@ import com.researchspace.model.Group;
 import com.researchspace.model.User;
 import com.researchspace.model.collection.AccessContext;
 import com.researchspace.model.collection.AccessContext.Operation;
-import com.researchspace.model.collection.CollectionDescription.Operator;
 import com.researchspace.model.collection.FilterExpression;
+import com.researchspace.model.collection.Operator;
 import com.researchspace.model.inventory.Instrument;
 import com.researchspace.model.inventory.InventoryRecord.InventorySharingMode;
 import com.researchspace.service.inventory.InstrumentReadAccess;
