+++ src/test/java/com/researchspace/dao/query/RuntimeFieldQueryTest.java
@@ -7,9 +7,9 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import com.researchspace.dao.query.RsqlCollectionQuery.Predicate;
 import com.researchspace.dao.query.RsqlCollectionQuery.Subquery;
 import com.researchspace.inventory.model.ApiV2InstrumentResource;
-import com.researchspace.model.collection.CollectionDescription.Operator;
 import com.researchspace.model.collection.CollectionQueryException;
 import com.researchspace.model.collection.FilterExpression;
+import com.researchspace.model.collection.Operator;
 import com.researchspace.model.collection.RelationshipReadAccess;
 import com.researchspace.model.collection.ResolvedRuntimeField;
 import com.researchspace.model.collection.RuntimeFieldBinding;
