+++ src/test/java/com/researchspace/model/collection/InMemoryCollectionQueryTest.java
@@ -3,9 +3,6 @@ package com.researchspace.model.collection;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import com.researchspace.model.collection.CollectionDescription.Field;
-import com.researchspace.model.collection.CollectionDescription.Operator;
-import com.researchspace.model.collection.CollectionDescription.Sort;
 import java.util.List;
 import org.junit.jupiter.api.Test;
